handle exception

+++ src/main/java/com/nvc/spring_boot/controller/UserController.java
@@ -2,6 +2,8 @@ package com.nvc.spring_boot.controller;
 
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,8 +11,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.nvc.spring_boot.entity.User;
+import com.nvc.spring_boot.domain.User;
 import com.nvc.spring_boot.service.UserService;
+import com.nvc.spring_boot.service.error.IdInvalidException;
+
 import org.springframework.web.bind.annotation.PutMapping;
 
 
@@ -22,30 +26,33 @@ public class UserController {
         this.userService = userService;
     }
 
-    @GetMapping("/user/list")
-    public List<User> getList() {
-        return userService.getList();
+    @GetMapping("/users/list")
+    public ResponseEntity<List<User>> getList() {
+        return ResponseEntity.ok(userService.getList());
     }
 
-    @GetMapping("/user/{id}")
-    public User findUser(@PathVariable("id") Long id) {
-        return userService.findUser(id);
+    @GetMapping("/users/{id}")
+    public ResponseEntity<User> findUser(@PathVariable("id") Long id) {
+        return ResponseEntity.ok(userService.findUser(id));
     }
 
-    @PutMapping("/user")
-    public User updateUser(@RequestBody User user) {
-        return userService.updateUser(user);
+    @PutMapping("/users")
+    public ResponseEntity<User> updateUser(@RequestBody User user) {
+        return ResponseEntity.ok(userService.updateUser(user));
     }
     
-    @PostMapping("/user")
-    public User createUser(@RequestBody User user) {
+    @PostMapping("/users")
+    public ResponseEntity<User> createUser(@RequestBody User user) {
         User newUser = userService.createUser(user);
-        return newUser;
+        return ResponseEntity.status(HttpStatus.CREATED).body(newUser);
     }
 
-    @DeleteMapping("/user/{id}")
-    public String deleteUser(@PathVariable("id") Long id) {
+    @DeleteMapping("/users/{id}")
+    public ResponseEntity<String> deleteUser(@PathVariable("id") Long id) throws IdInvalidException {
+        if (id >= 100) {
+            throw new IdInvalidException("Id must be less than 100");
+        }
         userService.deleteUser(id);
-        return "Delete User";
+        return ResponseEntity.ok("User deleted");
     }
 }
